made required fields and project
