+++ amplify/backend/backend-config.json
@@ -80,5 +80,11 @@
       ]
     }
   },
-  "storage": {}
+  "storage": {
+    "s3resumes1": {
+      "dependsOn": [],
+      "providerPlugin": "awscloudformation",
+      "service": "S3"
+    }
+  }
 }
