+++ amplify/backend/function/sparkinternshipwebe2c2bb07e2c2bb07PreSignup/src/email-filter-allowlist.js
@@ -2,13 +2,13 @@
  * @type {import('@types/aws-lambda').PreSignUpTriggerHandler}
  */
 exports.handler = async (event) => {
-  // allowed domains
-  const ald = process.env.DOMAINALLOWLIST.split(',').map((d) => d.trim());
+  // allowed domains using regex /^.+\.edu$/i
+  const ald = /^.+\.edu$/i;
 
   const { email } = event.request.userAttributes;
   const domain = email.substring(email.indexOf('@') + 1);
 
-  if (!ald.includes(domain)) {
+  if (!ald.test(domain)) {
     throw new Error(`Invalid email domain: ${domain}`);
   }
