+++ amplify/backend/function/sparkinternshipwebe2c2bb07e2c2bb07PreSignup/src/event.json
@@ -1,7 +1,7 @@
 {
   "request": {
     "userAttributes": {
-      "email": "testEmail@domain.com"
+      "email": "testEmail@domain.edu"
     }
   },
   "response": {}
