+++ amplify/backend/types/amplify-dependent-resources-ref.d.ts
@@ -25,5 +25,11 @@ export type AmplifyDependentResourcesAttributes = {
       "Name": "string",
       "Region": "string"
     }
+  },
+  "storage": {
+    "s3resumes1": {
+      "BucketName": "string",
+      "Region": "string"
+    }
   }
 }
