+++ amplify/team-provider-info.json
@@ -24,8 +24,11 @@
       "function": {
         "sparkinternshipwebe2c2bb07e2c2bb07PreSignup": {
           "deploymentBucketName": "amplify-sparkinternshipweb-dev-130617-deployment",
-          "s3Key": "amplify-builds/sparkinternshipwebe2c2bb07e2c2bb07PreSignup-4e2f466b343057696e66-build.zip"
+          "s3Key": "amplify-builds/sparkinternshipwebe2c2bb07e2c2bb07PreSignup-652b6e676b59634a2b35-build.zip"
         }
+      },
+      "storage": {
+        "s3resumes1": {}
       }
     }
   }
